# Improve parquet reading performance `~35-40%`