--- a/polars/polars-io/src/parquet/read.rs
+++ b/polars/polars-io/src/parquet/read.rs
@@ -6,16 +6,13 @@ use crate::prelude::*;
 use crate::RowCount;
 use arrow::io::parquet::read;
 use polars_core::prelude::*;
-use std::fs::File;
 use std::io::{Read, Seek};
-use std::path::Path;
 use std::sync::Arc;
 
 /// Read Apache parquet format into a DataFrame.
 #[must_use]
-pub struct ParquetReader<'a, R: Read + Seek> {
+pub struct ParquetReader<R: Read + Seek> {
     reader: R,
-    path: Option<&'a Path>,
     rechunk: bool,
     n_rows: Option<usize>,
     columns: Option<Vec<String>>,
@@ -24,18 +21,7 @@ pub struct ParquetReader<'a, R: Read + Seek> {
     row_count: Option<RowCount>,
 }
 
-impl<'a> ParquetReader<'a, File> {
-    /// Create a new [`ParquetReader`] from a known `path`.
-    /// Prefer `from_path` over `new` as that is faster.
-    pub fn from_path(path: &'a Path) -> Result<Self> {
-        let file = std::fs::File::open(path)?;
-        let mut out = Self::new(file);
-        out.path = Some(path);
-        Ok(out)
-    }
-}
-
-impl<R: MmapBytesReader> ParquetReader<'_, R> {
+impl<R: MmapBytesReader> ParquetReader<R> {
     #[cfg(feature = "lazy")]
     // todo! hoist to lazy crate
     pub fn finish_with_scan_ops(
@@ -51,7 +37,6 @@ impl<R: MmapBytesReader> ParquetReader<'_, R> {
         let rechunk = self.rechunk;
         read_parquet(
             self.reader,
-            self.path,
             self.n_rows.unwrap_or(usize::MAX),
             projection,
             &schema,
@@ -109,14 +94,11 @@ impl<R: MmapBytesReader> ParquetReader<'_, R> {
     }
 }
 
-impl<R: MmapBytesReader> SerReader<R> for ParquetReader<'_, R> {
-    /// Create a new [`ParquetReader`] for an existing `Reader`.
-    /// If reading from a file, prefer [`ParquetReader::from_path`], this
-    /// is faster.
+impl<R: MmapBytesReader> SerReader<R> for ParquetReader<R> {
+    /// Create a new [`ParquetReader`] from an existing `Reader`.
     fn new(reader: R) -> Self {
         ParquetReader {
             reader,
-            path: None,
             rechunk: false,
             n_rows: None,
             columns: None,
@@ -141,7 +123,6 @@ impl<R: MmapBytesReader> SerReader<R> for ParquetReader<'_, R> {
 
         read_parquet(
             self.reader,
-            self.path,
             self.n_rows.unwrap_or(usize::MAX),
             self.projection.as_deref(),
             &schema,