--- a/polars/polars-io/src/parquet/read_impl.rs
+++ b/polars/polars-io/src/parquet/read_impl.rs
@@ -15,10 +15,33 @@ use std::borrow::Cow;
 use std::convert::TryFrom;
 use std::io::Cursor;
 use std::ops::Deref;
+use std::path::Path;
 use std::sync::Arc;
 
-fn array_iter_to_series(iter: ArrayIter, field: &ArrowField) -> Result<Series> {
-    let chunks = iter.collect::<arrow::error::Result<Vec<_>>>()?;
+fn array_iter_to_series(
+    iter: ArrayIter,
+    field: &ArrowField,
+    num_rows: Option<usize>,
+) -> Result<Series> {
+    let mut total_count = 0;
+    let chunks = match num_rows {
+        None => iter.collect::<arrow::error::Result<Vec<_>>>()?,
+        Some(n) => {
+            let mut out = Vec::with_capacity(2);
+
+            for arr in iter {
+                let arr = arr?;
+                let len = arr.len();
+                out.push(arr);
+
+                total_count += len;
+                if total_count >= n {
+                    break;
+                }
+            }
+            out
+        }
+    };
     if chunks.is_empty() {
         let arr = new_empty_array(field.data_type.clone());
         Series::try_from((field.name.as_str(), arr))
@@ -29,19 +52,18 @@ fn array_iter_to_series(iter: ArrayIter, field: &ArrowField) -> Result<Series> {
 
 #[allow(clippy::too_many_arguments)]
 pub fn read_parquet<R: MmapBytesReader>(
-    reader: R,
+    mut reader: R,
+    path: Option<&Path>,
     limit: usize,
     projection: Option<&[usize]>,
     schema: &ArrowSchema,
     metadata: Option<FileMetaData>,
     predicate: Option<Arc<dyn PhysicalIoExpr>>,
     aggregate: Option<&[ScanAggregation]>,
-    parallel: bool,
+    mut parallel: bool,
     row_count: Option<RowCount>,
 ) -> Result<DataFrame> {
-    let reader = ReaderBytes::from(&reader);
-    let bytes = reader.deref();
-    let mut reader = Cursor::new(bytes);
+    let file = reader.to_file();
 
     let file_metadata = metadata
         .map(Ok)
@@ -51,6 +73,10 @@ pub fn read_parquet<R: MmapBytesReader>(
     let projection = projection
         .map(Cow::Borrowed)
         .unwrap_or_else(|| Cow::Owned((0usize..schema.fields.len()).collect::<Vec<_>>()));
+
+    if projection.len() == 1 || path.is_none() {
+        parallel = false;
+    }
 
     let mut dfs = Vec::with_capacity(row_group_len);
 
@@ -83,11 +109,13 @@ pub fn read_parquet<R: MmapBytesReader>(
 
         let chunk_size = md.num_rows() as usize;
         let columns = if parallel {
+            let file_name = path.unwrap();
+
             POOL.install(|| {
                 projection
                     .par_iter()
                     .map(|column_i| {
-                        let mut reader = Cursor::new(bytes);
+                        let mut reader = std::fs::File::open(file_name).unwrap();
                         let field = &schema.fields[*column_i];
                         let columns = read::read_columns(&mut reader, md.columns(), &field.name)?;
                         let iter = to_deserializer(
@@ -97,7 +125,11 @@ pub fn read_parquet<R: MmapBytesReader>(
                             Some(chunk_size),
                         )?;
 
-                        array_iter_to_series(iter, field)
+                        if remaining_rows < md.num_rows() {
+                            array_iter_to_series(iter, field, Some(remaining_rows))
+                        } else {
+                            array_iter_to_series(iter, field, None)
+                        }
                     })
                     .collect::<Result<Vec<_>>>()
             })?
@@ -110,7 +142,11 @@ pub fn read_parquet<R: MmapBytesReader>(
                     let iter =
                         to_deserializer(columns, field.clone(), remaining_rows, Some(chunk_size))?;
 
-                    array_iter_to_series(iter, field)
+                    if remaining_rows < md.num_rows() {
+                        array_iter_to_series(iter, field, Some(remaining_rows))
+                    } else {
+                        array_iter_to_series(iter, field, None)
+                    }
                 })
                 .collect::<Result<Vec<_>>>()?
         };