--- a/polars/polars-lazy/src/physical_plan/executors/scan/parquet.rs
+++ b/polars/polars-lazy/src/physical_plan/executors/scan/parquet.rs
@@ -36,7 +36,7 @@ impl ParquetExec {
             &self.aggregate,
         );
 
-        ParquetReader::from_path(self.path.as_path())?
+        ParquetReader::new(file)
             .with_n_rows(n_rows)
             .read_parallel(self.options.parallel)
             .with_row_count(std::mem::take(&mut self.options.row_count))