@alexander-beedie 2022-06-28
That's quite a gain - especially given that it was hardly slow to start with :)

<img src="https://user-images.githubusercontent.com/2613171/176114942-fb3863c8-1271-4685-b7d2-a2c52045e279.jpg" width="25%">