@thobai 2022-07-01
@ritchie46 This sounds very promising! Did that already make it into the recent python version of polars (0.13.51)? I just tested with that version and I can unfortunately not really see a difference in performance for my datasets and queries. But that could also be because the majority of time is spent somewhere else and reading wasn't a bit fraction of the execution time.